% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color_space.r
\docType{data}
\name{hcl_presets}
\alias{hcl_presets}
\title{A list of color space presets as defined in \url{http://tools.medialab.sciences-po.fr/iwanthue/js/presets.js}.}
\format{\preformatted{List of 1
 $ full:List of 3
  ..$ H: num [1:2] 0 360
  ..$ C: num [1:2] 0 100
  ..$ L: num [1:2] 0 100
}}
\usage{
hcl_presets
}
\description{
A list of color space presets as defined in \url{http://tools.medialab.sciences-po.fr/iwanthue/js/presets.js}.
}
\keyword{datasets}

