% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwanthue.r
\name{inwanthueInstance}
\alias{inwanthueInstance}
\title{IWantHue instance accessor}
\usage{
inwanthueInstance(seed, force_init)
}
\arguments{
\item{seed}{The seed for the rng}

\item{force_init}{If true, force recreation of the instance}
}
\description{
IWantHue instance accessor
}
