% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iwanthue.r
\name{iwanthue}
\alias{iwanthue}
\title{Generate a new iwanthue palette}
\usage{
iwanthue(n = 8, color_space = list(L = c(0, 100), C = c(0, 100), H = c(0,
  360)), mode = "ga", ...)
}
\arguments{
\item{n}{number of colors in the palette.}

\item{color_space}{the upper and lower limts of the hcl colors to put in the palette. Expects a list of three numerics (See \code{\link{hcl_presets}}).}

\item{mode}{the algorithm to use.}
}
\value{
a new rgb hex color palette of length \code{n}.
}
\description{
Generate a new iwanthue palette
}

