% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rwantshue.r
\docType{package}
\name{rwantshue}
\alias{rwantshue}
\alias{rwantshue-package}
\title{rwantshue is a R adaptor for the "I want hue" color palette generator}
\description{
rwantshue is a R adaptor for the "I want hue" color palette generator
}

