% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwanthue.r
\docType{data}
\name{singletonEnv}
\alias{singletonEnv}
\title{The environemnt that holds a IWantHue instance}
\format{An object of class \code{environment} of length 0.}
\usage{
singletonEnv
}
\description{
The environemnt that holds a IWantHue instance
}
\keyword{datasets}
