% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hcl.r
\docType{data}
\name{hcl_presets}
\alias{hcl_presets}
\title{A list of color space presets as defined in \url{http://tools.medialab.sciences-po.fr/iwanthue/js/presets.js}.}
\format{\preformatted{List of 16
 $ fancy_light :List of 3
  ..$ : num [1:2] 0 360
  ..$ : num [1:2] 0.4 1.2
  ..$ : num [1:2] 1 1.5
 $ fancy_dark  :List of 3
  ..$ : num [1:2] 0 360
  ..$ : num [1:2] 0.2 1.2
  ..$ : num [1:2] 0.1 0.6
 $ shades      :List of 3
  ..$ : num [1:2] 0 240
  ..$ : num [1:2] 0 0.4
  ..$ : num [1:2] 0 1.5
 $ tarnish     :List of 3
  ..$ : num [1:2] 0 360
  ..$ : num [1:2] 0 0.4
  ..$ : num [1:2] 0.4 1.1
 $ pastel      :List of 3
  ..$ : num [1:2] 0 360
  ..$ : num [1:2] 0 0.9
  ..$ : num [1:2] 1 1.5
 $ pimp        :List of 3
  ..$ : num [1:2] 0 360
  ..$ : num [1:2] 0.9 3
  ..$ : num [1:2] 0.4 1
 $ intense     :List of 3
  ..$ : num [1:2] 0 360
  ..$ : num [1:2] 0.6 3
  ..$ : num [1:2] 0.2 1.1
 $ fluo        :List of 3
  ..$ : num [1:2] 0 300
  ..$ : num [1:2] 1 3
  ..$ : num [1:2] 1.1 1.5
 $ red_roses   :List of 3
  ..$ : num [1:2] 20 330
  ..$ : num [1:2] 0.3 3
  ..$ : num [1:2] 0.5 1.5
 $ ochre_sand  :List of 3
  ..$ : num [1:2] 20 60
  ..$ : num [1:2] 0.3 1.6
  ..$ : num [1:2] 0.5 1.5
 $ yellow_lime :List of 3
  ..$ : num [1:2] 60 90
  ..$ : num [1:2] 0.3 3
  ..$ : num [1:2] 0.5 1.5
 $ green_mint  :List of 3
  ..$ : num [1:2] 90 150
  ..$ : num [1:2] 0.3 3
  ..$ : num [1:2] 0.5 1.5
 $ ice_cube    :List of 3
  ..$ : num [1:2] 150 200
  ..$ : num [1:2] 0 3
  ..$ : num [1:2] 0.5 1.5
 $ blue_ocean  :List of 3
  ..$ : num [1:2] 220 260
  ..$ : num [1:2] 0.2 2.5
  ..$ : num [1:2] 0 0.8
 $ indigo_night:List of 3
  ..$ : num [1:2] 260 290
  ..$ : num [1:2] 1.2 3
  ..$ : num [1:2] 0.5 1.5
 $ purple_white:List of 3
  ..$ : num [1:2] 290 330
  ..$ : num [1:2] 0 3
  ..$ : num [1:2] 0 0.6
}}
\usage{
hcl_presets
}
\description{
A list of color space presets as defined in \url{http://tools.medialab.sciences-po.fr/iwanthue/js/presets.js}.
}
\keyword{datasets}

