% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl.r
\docType{data}
\name{hcl_presets}
\alias{hcl_presets}
\title{A list of color space presets as defined in \url{http://tools.medialab.sciences-po.fr/iwanthue/js/presets.js}.}
\format{An object of class \code{list} of length 19.}
\usage{
hcl_presets
}
\description{
A list of color space presets as defined in \url{http://tools.medialab.sciences-po.fr/iwanthue/js/presets.js}.
}
\keyword{datasets}
