% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iwanthue.r
\docType{class}
\name{IWantHue-class}
\alias{IWantHue}
\alias{IWantHue-class}
\title{IWantHue palette generator.}
\description{
IWantHue palette generator.
}
\section{Fields}{

\describe{
\item{\code{v8}}{The V8 context.}
}}
\section{Methods}{

\describe{
\item{\code{hex(...)}}{Generate a vector of colors in hex format}

\item{\code{rgb(...)}}{Generate a matrix of colors in rgb format}
}}

