% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwanthue.r
\name{iwanthue}
\alias{iwanthue}
\title{Get the \linkS4class{IWantHue} instance.}
\usage{
iwanthue(seed = NULL, force_init = FALSE)
}
\arguments{
\item{seed}{The seed for the rng}

\item{force_init}{If true, force recreation of the instance}
}
\description{
Get the \linkS4class{IWantHue} instance.
}
