% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iwanthue_js.r
\docType{data}
\name{singletonEnv}
\alias{singletonEnv}
\title{The environemnt that holds a IWantHue instance}
\format{\preformatted{<environment: 0x7fe6a6aae1f0> 
}}
\usage{
singletonEnv
}
\description{
The environemnt that holds a IWantHue instance
}
\keyword{datasets}

