% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iwanthue.r
\docType{data}
\name{singletonEnv}
\alias{singletonEnv}
\title{The environemnt that holds a IWantHue instance}
\format{\preformatted{<environment: 0x7fd5a3576b10> 
}}
\usage{
singletonEnv
}
\description{
The environemnt that holds a IWantHue instance
}
\keyword{datasets}

